Add guide for comparing Pony to language X section
This guide gives advice for how to write a chapter for the book, with
examples of features of Pony to compare and contrast with other
languages.

The goal is to encourage writers to compare interesting features of
the two languages that do or do not overlap, rather than focusing on
surface things like the syntax of conditional statements or showing
non-idomatic examples.